Изменены include, добавлены пробелы

+++ Expression/arrayexpression.cpp
@@ -1,26 +1,27 @@
-#include "arrayexpression.h"
-#include "../Value/arrayvalue.h"
+#include <Expression/arrayexpression.h>
+#include <Value/arrayvalue.h>
+
 
 using namespace SlavaScript::lang;
 
 
 ArrayExpression::ArrayExpression(std::vector<Expression*> elements) : elements(elements) {}
 
-std::shared_ptr<Value> ArrayExpression::eval(){
+std::shared_ptr<Value> ArrayExpression::eval() {
     ArrayValue arr(elements.size());
-    for(int i = 0; i < elements.size(); ++i){
+    for (size_t i = 0; i < elements.size(); ++i) {
         arr.set(i, elements[i] -> eval());
     }
     SH_RET(ArrayValue, arr);
 }
 
-Expressions ArrayExpression::type() const{
+Expressions ArrayExpression::type() const {
     return Expressions::ArrayExpression;
 }
 
-ArrayExpression::operator std::string(){
+ArrayExpression::operator std::string() {
     std::string result = "[";
-    for(int i = 0; i < elements.size(); ++i){
+    for (size_t i = 0; i < elements.size(); ++i) {
         result += std::string(*(elements[i]));
         if (i < elements.size() - 1) result += ", ";
     }
@@ -28,13 +29,13 @@ ArrayExpression::operator std::string(){
     return result;
 }
 
-ArrayExpression::~ArrayExpression(){
-    for(int i = 0; i < elements.size(); ++i){
+ArrayExpression::~ArrayExpression() {
+    for (size_t i = 0; i < elements.size(); ++i) {
         delete elements[i];
         elements[i] = nullptr;
     }
 }
 
-void ArrayExpression::accept(Visitor* visitor){
+void ArrayExpression::accept(Visitor* visitor) {
     visitor -> visit(this);
 }
