+++ Parser/parseerrors.cpp
@@ -1,24 +1,25 @@
-#include "parseerrors.h"
+#include <Parser/parseerrors.h>
+
 
 using namespace SlavaScript::lang;
 using SlavaScript::exceptions::ParseException;
 
 ParseErrors::ParseErrors() {}
 
-void ParseErrors::clear(){
+void ParseErrors::clear() {
     errors.clear();
 }
 
-void ParseErrors::add(ParseException ex, int line){
+void ParseErrors::add(ParseException ex, int line) {
     errors.push_back(ParseError(line, ex));
 }
 
-bool ParseErrors::hasError(){
+bool ParseErrors::hasError() {
     return !errors.empty();
 }
 
-ParseErrors::operator std::string(){
+ParseErrors::operator std::string() {
     std::string result;
-    for(int i = 0; i < errors.size(); ++i) result += std::string(errors[i]) + "\n";
+    for (size_t i = 0; i < errors.size(); ++i) result += std::string(errors[i]) + "\n";
     return result;
 }
