+++ Run/path.cpp
@@ -1,7 +1,8 @@
-#include "path.h"
-#include "../Lib/filesystem.h"
-#include "../Lib/names.h"
-#include "../Value/stringvalue.h"
+#include <Lib/filesystem.h>
+#include <Lib/names.h>
+#include <Run/path.h>
+#include <Value/stringvalue.h>
+
 
 using namespace SlavaScript::lang;
 
@@ -9,36 +10,36 @@ std::vector<std::string> Path::pathes = {};
 bool Path::import = true;
 std::shared_ptr<ArrayValue> Path::arguments = nullptr;
 
-void Path::setPath(std::string path){
+void Path::setPath(std::string path) {
     pathes.push_back(path);
 }
 
-std::string Path::getPath(){
+std::string Path::getPath() {
     return pathes.back();
 }
 
-void Path::popPath(){
+void Path::popPath() {
     pathes.pop_back();
 }
 
-bool Path::getImpoted(){
+bool Path::getImpoted() {
     return import;
 }
 
-void Path::setImpoted(bool imp){
+void Path::setImpoted(bool imp) {
     import = imp;
 }
 
-void Path::initContainers(){
+void Path::initContainers() {
     Names::init();
 }
 
-void Path::setCommandArguments(std::vector<std::string> argv){
+void Path::setCommandArguments(std::vector<std::string> argv) {
     arguments = SHARE(ArrayValue, argv.size());
-    for(int i = 0; i < argv.size(); ++i) arguments -> set(i, SHARE(StringValue, argv[i]));
+    for (size_t i = 0; i < argv.size(); ++i) arguments -> set(i, SHARE(StringValue, argv[i]));
     FS::setApplicationPath(argv[0]);
 }
 
-std::shared_ptr<ArrayValue> Path::getCommandArguments(){
+std::shared_ptr<ArrayValue> Path::getCommandArguments() {
     return arguments;
 }
