+++ Statement/switchstatement.cpp
@@ -1,44 +1,45 @@
-#include "switchstatement.h"
-#include "breakstatement.h"
-#include "continuestatement.h"
+#include <Statement/breakstatement.h>
+#include <Statement/continuestatement.h>
+#include <Statement/switchstatement.h>
+
 
 using namespace SlavaScript::lang;
 
 SwitchStatement::SwitchStatement(Expression* start, std::vector<std::pair<Expression*, Statement*>> body, Statement* defaultCase) : start(start), body(body), defaultCase(defaultCase) {}
 
-void SwitchStatement::execute(){
+void SwitchStatement::execute() {
     bool trueCase = false;
-    try{
-        for(int i = 0; i < body.size(); ++i){
-            if (*(start -> eval()) == *(body[i].first -> eval())){
+    try {
+        for (size_t i = 0; i < body.size(); ++i) {
+            if (*(start -> eval()) == *(body[i].first -> eval())) {
                 trueCase = true;
-                try{
+                try {
                     body[i].second -> execute();
-                } catch(ContinueStatement& cs){
+                } catch (ContinueStatement& cs) {
                     // continue;
                 }
             }
         }
         if (!trueCase && defaultCase != nullptr) defaultCase -> execute();
-    } catch(BreakStatement& bs){
+    } catch (BreakStatement& bs) {
         // break;
     }
 }
 
-Statements SwitchStatement::type() const{
+Statements SwitchStatement::type() const {
     return Statements::SwitchStatement;
 }
 
-SwitchStatement::operator std::string(){
+SwitchStatement::operator std::string() {
     std::string result = "switch " + std::string(*start) + " {\n";
-    for (unsigned i = 0; i < body.size(); ++i){
+    for (size_t i = 0; i < body.size(); ++i) {
         result += "case ";
         result += std::string(*body[i].first);
         result += " : ";
         result += std::string(*body[i].second);
         result += "\n";
     }
-    if (defaultCase != nullptr){
+    if (defaultCase != nullptr) {
         result += "default : ";
         result += std::string(*defaultCase);
         result += "\n";
@@ -47,10 +48,10 @@ SwitchStatement::operator std::string(){
     return result;
 }
 
-SwitchStatement::~SwitchStatement(){
+SwitchStatement::~SwitchStatement() {
     delete start;
     start = nullptr;
-    for(unsigned i = 0; i < body.size(); ++i){
+    for (size_t i = 0; i < body.size(); ++i) {
         delete body[i].first;
         body[i].first = nullptr;
         delete body[i].second;
@@ -60,6 +61,6 @@ SwitchStatement::~SwitchStatement(){
     defaultCase = nullptr;
 }
 
-void SwitchStatement::accept(Visitor* visitor){
+void SwitchStatement::accept(Visitor* visitor) {
     visitor -> visit(this);
 }
